[fiz] Исправляет замечания после ревью
