+++ .gitignore
@@ -1,4 +1,6 @@
 *.iml
+.idea
+build
 .gradle
 /local.properties
 /.idea/caches
