+++ app/src/main/java/com/example/playlistmaker/trackReciclerView/Track.kt
@@ -1,0 +1,8 @@
+package com.example.playlistmaker.trackReciclerView
+
+data class Track(
+    val trackName: String,
+    val trackTime: String,
+    val artistName: String,
+    val artWorkUrl100: String
+)
