+++ app/src/main/java/com/example/playlistmaker/trackReciclerView/TrackAdapter.kt
@@ -1,0 +1,25 @@
+package com.example.playlistmaker.trackReciclerView
+
+import android.view.LayoutInflater
+import android.view.ViewGroup
+import androidx.recyclerview.widget.RecyclerView
+import com.example.playlistmaker.R
+
+class TrackAdapter(
+    private val news: List<Track>
+) : RecyclerView.Adapter<TrackViewHolder>() {
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): TrackViewHolder {
+        val view =
+            LayoutInflater.from(parent.context).inflate(R.layout.track_view, parent, false)
+        return TrackViewHolder(view)
+    }
+
+    override fun onBindViewHolder(holder: TrackViewHolder, position: Int) {
+        holder.bind(news[position])
+    }
+
+    override fun getItemCount(): Int {
+        return news.size
+    }
+}
