+++ app/src/main/res/layout/track_view.xml
@@ -9,14 +9,14 @@
 
     <ImageView
         android:id="@+id/trackPreview"
-        android:layout_width="45dp"
-        android:layout_height="45dp"
+        android:layout_width="44dp"
+        android:layout_height="44dp"
+        android:background="@drawable/rounded_corners_xs"
+        android:clipToOutline="true"
         android:contentDescription="@string/track_preview"
         android:scaleType="centerCrop"
-        android:src="@drawable/placeholder"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        tools:src="@drawable/placeholder" />
+        app:layout_constraintTop_toTopOf="parent" />
 
     <TextView
         android:id="@+id/trackName"
